{"schema":"decodex/commit/1","summary":"Extract Swift export encoder bridge owner","authority":"manual"}
